refactor(Screening) : 자바 문법에 따라 수정, 패키지 네이밍 schedule -> screening 변경
